Register welcome event only on debug mode
